Colour defs

+++ PoE-TradeUI/PoE-TradeUI.Core/Defs/ColourDef.cs
@@ -1,5 +1,5 @@
 ﻿namespace PoE_TradeUI.Core.Defs {
     public class ColourDef {
-        public int[] Rgb;
+        public byte[] Rgb;
     }
 }
